Ramin Yazdani | Protein Homology Modeling | main | feat(data): Add target protein sequence
Add target protein sequence in FASTA format.

The target_sequence.fasta file contains the amino acid sequence of the protein
to be modeled. This is one of the two essential inputs for homology modeling.

FASTA format:
- Header line with sequence identifier
- Amino acid sequence in single-letter code

This sequence will be aligned to the template structure to build the homology model.

Verified: target_sequence.fasta added with valid FASTA format.